rename group to channel within community
