+++ frontend/app/src/stores/community.ts
@@ -6,7 +6,7 @@ function createDummyCommunity(
     name = `Community name ${id}`,
     url = "../assets/unknownUserAvatar.svg",
     memberCount = 2000,
-    groupCount = 15,
+    channelCount = 15,
     unreadCount = 0
 ) {
     return {
@@ -15,17 +15,17 @@ function createDummyCommunity(
         description:
             "This is an awsome community with lots of interesting things to see and do. Not too much racism. Not financial advice. HODL.",
         memberCount,
-        groupCount,
+        channelCount,
         unreadCount,
         blobUrl: url,
     };
 }
 
-function createDummyCommunityGroup(id: string) {
+function createDummyCommunityChannel(id: string) {
     return {
-        name: `Group ${id}`,
+        name: `Channel ${id}`,
         description:
-            "This is a nice groups that belongs to this community. It might have some _markdown_ or even some `code samples`.",
+            "This is a nice channel that belongs to this community. It might have some _markdown_ or even some `code samples`.",
     };
 }
 
@@ -53,19 +53,19 @@ export const dummyCommunities = writable<Community[]>([
     createDummyCommunity("14"),
 ]);
 
-export const dummyCommunityGroups = writable<{ name: string; description: string }[]>([
-    createDummyCommunityGroup("1"),
-    createDummyCommunityGroup("2"),
-    createDummyCommunityGroup("3"),
-    createDummyCommunityGroup("4"),
-    createDummyCommunityGroup("5"),
-    createDummyCommunityGroup("6"),
-    createDummyCommunityGroup("7"),
-    createDummyCommunityGroup("8"),
-    createDummyCommunityGroup("9"),
-    createDummyCommunityGroup("10"),
-    createDummyCommunityGroup("11"),
-    createDummyCommunityGroup("12"),
-    createDummyCommunityGroup("13"),
-    createDummyCommunityGroup("14"),
+export const dummyCommunityChannels = writable<{ name: string; description: string }[]>([
+    createDummyCommunityChannel("1"),
+    createDummyCommunityChannel("2"),
+    createDummyCommunityChannel("3"),
+    createDummyCommunityChannel("4"),
+    createDummyCommunityChannel("5"),
+    createDummyCommunityChannel("6"),
+    createDummyCommunityChannel("7"),
+    createDummyCommunityChannel("8"),
+    createDummyCommunityChannel("9"),
+    createDummyCommunityChannel("10"),
+    createDummyCommunityChannel("11"),
+    createDummyCommunityChannel("12"),
+    createDummyCommunityChannel("13"),
+    createDummyCommunityChannel("14"),
 ]);
